requirements.txt
asgiref==3.9.0
asttokens==3.0.0
black==25.1.0
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
coverage==7.9.2
decorator==5.2.1
distlib==0.3.9
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
executing==2.2.0
filelock==3.18.0
flake8==7.3.0
identify==2.6.12
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
PyYAML==6.0.2
rich==14.0.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
virtualenv==20.31.2
wcwidth==0.2.13
mysqlclient==2.2.4
pymysql==1.1.1
python-dotenv==1.0.1
cryptography>=42.0
boto3==1.39.9
django-storages==1.14.6
